# robin/openapi.yaml
openapi: "3.1.0"
info:
  title: Robin API
  version: "0.3.0"
  description: REST API for Robin.OS — a personal knowledge management system.
servers:
  - url: http://localhost:3000
    description: Local development
security:
  - cookieAuth: []
...
paths:
...
  /threads/{id}:
    get:
      operationId: getThread
      summary: Get a thread by ID (includes wiki content)
      tags: [Threads]
      parameters:
        - $ref: "#/components/parameters/Id"
      responses:
        "200":
          description: Thread with wiki content
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/threadWithWikiResponseSchema"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/errorResponseSchema"
    put:
      operationId: updateThread
      summary: Update a thread
      tags: [Threads]
      parameters:
        - $ref: "#/components/parameters/Id"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/updateThreadBodySchema"
      responses:
        "200":
          description: Updated thread
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/threadResponseSchema"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/errorResponseSchema"
  /threads/{id}/regenerate:
    post:
      operationId: regenerateThread
      summary: Trigger thread wiki regeneration
      tags: [Threads]
      parameters:
        - $ref: "#/components/parameters/Id"
      responses:
        "202":
          description: Job queued
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/queuedResponseSchema"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/errorResponseSchema"
  /threads/{targetId}/merge:
    post:
      operationId: mergeThreads
      summary: Merge threads (not implemented)
      tags: [Threads]
      parameters:
        - name: targetId
          in: path
          required: true
          schema:
            type: string
      responses:
        "501":
          description: Not implemented
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/errorResponseSchema"

  # ── Vaults ─────────────────────────────────────────────────────────
...
  /users/profile:
    get:
      operationId: getUserProfile
      summary: Get current user's profile
      tags: [Users]
      responses:
        "200":
          description: User profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/userProfileResponseSchema"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/errorResponseSchema"
...
  /users/keypair:
    get:
      operationId: getUserKeypair
      summary: Get user's Ed25519 keypair
      tags: [Users]
      responses:
        "200":
          description: Keypair details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/keypairResponseSchema"
        "404":
          description: No keypair
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/errorResponseSchema"
...
  /users/regenerate-mcp:
    post:
      operationId: regenerateMcpEndpoint
      summary: Regenerate MCP endpoint URL
      tags: [Users]
      responses:
        "200":
          description: New MCP endpoint
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/mcpEndpointResponseSchema"
        "400":
          description: No keypair
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/errorResponseSchema"
...
components:
...
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: better-auth.session_token
      description: Session cookie set by better-auth after sign-in
    hmacAuth:
      type: apiKey
      in: header
      name: X-Signature
      description: HMAC-SHA256 signature of request body
    mcpToken:
      type: apiKey
      in: query
      name: token
      description: JWT token for MCP access

  # ── Shared Parameters ──────────────────────────────────────────────
  parameters:
    Id:
      name: id
      in: path
      required: true
      description: Resource ID or lookup key
      schema:
        type: string
...
  schemas:
...
    errorResponseSchema:
      type: object
      properties:
        error:
          type: string
        fields: {}
      required: [error]
      additionalProperties: false
...
    queuedResponseSchema:
      type: object
      properties:
        status:
          type: string
          enum: [queued]
        jobId:
          type: string
      required: [status, jobId]
      additionalProperties: false

    # ── Entries ─────────────────────────────────────────────────────
...
    updateThreadBodySchema:
      type: object
      properties:
        name:
          type: string
        type:
          type: string
        prompt:
          type: string
      additionalProperties: false
    threadResponseSchema:
      type: object
      properties:
        id:
          type: string
          minLength: 1
        lookupKey:
          type: string
          minLength: 1
        userId:
          type: string
        slug:
          type: string
        name:
          type: string
        type:
          type: string
        prompt:
          type: string
        state:
          type: string
          enum: [PENDING, RESOLVED, LINKING, DIRTY]
        repoPath:
          type: string
        vaultId:
          type: string
          nullable: true
        lastRebuiltAt:
          type: string
          format: date-time
          nullable: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        noteCount:
          type: number
          default: 0
        lastUpdated:
          type: string
      required: [id, lookupKey, userId, slug, name, type, prompt, state, repoPath, vaultId, lastRebuiltAt, createdAt, updatedAt, lastUpdated]
      additionalProperties: false
    threadWithWikiResponseSchema:
      type: object
      properties:
        id:
          type: string
          minLength: 1
        lookupKey:
          type: string
          minLength: 1
        userId:
          type: string
        slug:
          type: string
        name:
          type: string
        type:
          type: string
        prompt:
          type: string
        state:
          type: string
          enum: [PENDING, RESOLVED, LINKING, DIRTY]
        repoPath:
          type: string
        vaultId:
          type: string
          nullable: true
        lastRebuiltAt:
          type: string
          format: date-time
          nullable: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        noteCount:
          type: number
          default: 0
        lastUpdated:
          type: string
        wikiContent:
          type: string
      required: [id, lookupKey, userId, slug, name, type, prompt, state, repoPath, vaultId, lastRebuiltAt, createdAt, updatedAt, lastUpdated, wikiContent]
      additionalProperties: false
...
    userProfileResponseSchema:
      type: object
      properties:
        id:
          type: string
        email:
          type: string
        name:
          type: string
          nullable: true
        mcpEndpointUrl:
          type: string
        apiKeyHint:
          type: string
        onboardedAt:
          type: string
          nullable: true
      required: [id, email, name, mcpEndpointUrl, apiKeyHint, onboardedAt]
      additionalProperties: false
...
    keypairResponseSchema:
      type: object
      properties:
        algorithm:
          type: string
        publicKey:
          type: string
        privateKey:
          type: string
      required: [algorithm, publicKey, privateKey]
      additionalProperties: false
    mcpEndpointResponseSchema:
      type: object
      properties:
        mcpEndpointUrl:
          type: string
      required: [mcpEndpointUrl]
      additionalProperties: false
...